Make `json_pure` an alias of `json_gem`
`json_pure` is no more, and all rubies since ages ship with an
embeded version of the `json` gem.

In the future it would be good to deprecate the `json_pure` adapter.
