+++ lib/multi_json/adapters/json_pure.rb
@@ -1,11 +1,7 @@
-require "json/pure"
-require_relative "json_common"
+require_relative "json_gem"
 
 module MultiJson
   module Adapters
-    # Use JSON pure to dump/load.
-    class JsonPure < JsonCommon
-      ParseError = ::JSON::ParserError
-    end
+    JsonPure = JsonGem
   end
 end
